Optimise size of model structs (#2464)
This commit reduces the sizes of many integer fields, or converts data to enums when appropriate:

- Reduce the size of rate_limit fields
- Reduce the size of channel position
- Reduce size of role position
- Use an enum for `auto_archive_duration`
- `DefaultReaction` -> `ForumEmoji`
- Swap `AfkTimeout` for enum
- Reduce `Sticker::sort_value` to u16
- Reduce size of `min`/`max_values`
- Reduce size of `max_uses` and `max_age`
- Reduce size of `ActivityParty::size`
- Reduce size of `Ready::version`
- Reduce `InputText::min/max_length`
- Reduce `size`/`width`/`height` -- Discord filesize limits mean we won't hit the size limit, and therefore won't hit height/width limits, plus discord rejects 0 sized files so we can use `NonZero*`
- Reduce size of `mention_total_limit`
- Reduce size of `RoleSubscriptionData::total_months_subscribed` -- This gives us ~5000 years of role subscription
- Fix Change sizes

+++ src/builder/create_invite.rs
@@ -73,9 +73,9 @@ use crate::model::prelude::*;
 #[must_use]
 pub struct CreateInvite<'a> {
     #[serde(skip_serializing_if = "Option::is_none")]
-    max_age: Option<u64>,
+    max_age: Option<u32>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    max_uses: Option<u64>,
+    max_uses: Option<u8>,
     #[serde(skip_serializing_if = "Option::is_none")]
     temporary: Option<bool>,
     #[serde(skip_serializing_if = "Option::is_none")]
@@ -124,7 +124,7 @@ impl<'a> CreateInvite<'a> {
     /// # Ok(())
     /// # }
     /// ```
-    pub fn max_age(mut self, max_age: u64) -> Self {
+    pub fn max_age(mut self, max_age: u32) -> Self {
         self.max_age = Some(max_age);
         self
     }
@@ -156,7 +156,7 @@ impl<'a> CreateInvite<'a> {
     /// # Ok(())
     /// # }
     /// ```
-    pub fn max_uses(mut self, max_uses: u64) -> Self {
+    pub fn max_uses(mut self, max_uses: u8) -> Self {
         self.max_uses = Some(max_uses);
         self
     }
